[COMPILER] * BUG: Do not create debugger information for code generated by form files.   Otherwise the debugger is confused.
git-svn-id: svn://localhost/gambas/trunk@3779 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/share/gb_buffer.h
@@ -36,7 +36,7 @@ typedef
 void BUFFER_create(void *p_data);
 void BUFFER_delete(void *p_data);
 bool BUFFER_load_file(void *p_data, const char *name);
-offset_t BUFFER_add(void *p_data, const void *string, size_t len);
+offset_t BUFFER_add(void *p_data, const void *string, int len);
 bool BUFFER_need(void *p_data, size_t size);
 
 #define DATA_TO_BUFFER(_data) ((BUFFER *)(_data) - 1)
